[o11y] Refactor Streaming Tail Worker to facilitate downstream usage
Also fixes promise continuations to ensure that the tail worker can process
events properly before being marked as completed.
