+++ src/workerd/api/tail-worker-test.wd-test
@@ -34,8 +34,9 @@ const unitTests :Workerd.Config = (
     (name = "alarms", worker = .alarmsWorker),
     (name = "hiber", worker = .hiberWorker),
     (name = "TEST_TMPDIR", disk = (writable = true)),
-    # Receives trace events from STW, used to check that STW produces trace events properly.
-    (name = "receiver", worker = .logReceiver, ),
+    # Receives trace events from STW, used to check that STW produces trace events properly –
+    # disabled until instrumentation for STW itself is supported.
+    # (name = "receiver", worker = .logReceiver, ),
     # Dummy legacy tail worker (gets traces from alarms worker and produces trace for main tracer)
     (name = "legacy", worker = .logLegacy, ),
     # Unified tail worker with tests
@@ -94,7 +95,7 @@ const logWorker :Workerd.Worker = (
   ],
   compatibilityDate = "2024-10-14",
   compatibilityFlags = ["experimental", "nodejs_compat"],
-  tails = ["receiver"],
+  # tails = ["receiver"],
 );
 
 const logReceiver :Workerd.Worker = (
